test(textarea): migrate item
