+++ core/src/components/textarea/test/item/textarea.e2e.ts-snapshots/textarea-inset-list-no-fill-ios-ltr-Mobile-Chrome-linux.png
