+++ core/src/components/textarea/test/item/textarea.e2e.ts-snapshots/textarea-inset-list-no-fill-md-ltr-Mobile-Chrome-linux.png
