+++ core/src/components/textarea/test/item/textarea.e2e.ts-snapshots/textarea-inset-list-no-fill-md-rtl-Mobile-Firefox-linux.png
