+++ core/src/components/textarea/test/item/textarea.e2e.ts-snapshots/textarea-list-no-fill-ios-ltr-Mobile-Safari-linux.png
