+++ core/src/components/textarea/test/item/textarea.e2e.ts-snapshots/textarea-list-no-fill-md-ltr-Mobile-Firefox-linux.png
